Turn articles section into an h-feed
Rather than turning the entire homepage into a monolithic feed only part
of the page is turned into an h-feed, a “partial feed”. Every entry has
also been provided with a matching publication date, as accurate as the
sources made possible.

The default stylesheet is updated so timestamps will always stay
together on smaller screens, never moving only the day or the year to a
new line.